Fixing wrong imports due to changes in the internal interface
