+++ services/web/server/src/simcore_service_webserver/login/config.py
@@ -3,13 +3,17 @@
     - config-file schema
     - settings
 """
-from typing import Optional, Dict
-from aiohttp.web import Application
+from typing import Dict, Optional
 
 import trafaret as T
+from aiohttp.web import Application
 from pydantic import BaseSettings
 
+from servicelib.application_keys import APP_CONFIG_KEY
+
+from ..email_config import CONFIG_SECTION_NAME as SMTP_SECTION
 from .cfg import DEFAULTS
+from .storage import AsyncpgStorage
 
 CONFIG_SECTION_NAME = "login"
 
@@ -41,7 +45,6 @@ class LoginSettings(BaseSettings):
 
 
 def get_login_config(app: Application) -> Dict:
-    from servicelib.application_keys import APP_CONFIG_KEY
 
     cfg = app[APP_CONFIG_KEY].get(CONFIG_SECTION_NAME, dict())
     return cfg
@@ -54,3 +57,27 @@ def assert_valid_config(app: Application) -> Dict:
     cfg = get_login_config(app)
     _settings = LoginSettings(**cfg)
     return cfg
+
+
+def create_login_internal_config(app: Application, storage: AsyncpgStorage) -> Dict:
+    """
+    Creates compatible config to update login.cfg.cfg object
+    """
+    login_cfg = app[APP_CONFIG_KEY].get(CONFIG_SECTION_NAME, {})  # optional!
+    smtp_cfg = app[APP_CONFIG_KEY][SMTP_SECTION]
+
+    config = {"APP": app, "STORAGE": storage}
+
+    def _fmt(val):
+        if isinstance(val, str):
+            if val.strip().lower() in ["null", "none", ""]:
+                return None
+        return val
+
+    for key, value in login_cfg.items():
+        config[key.upper()] = _fmt(value)
+
+    for key, value in smtp_cfg.items():
+        config["SMTP_{}".format(key.upper())] = _fmt(value)
+
+    return config
