+++ services/web/server/tests/unit/with_dbs/fast/test_change_email.py
@@ -7,8 +7,7 @@ from yarl import URL
 
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import LoggedUser, NewUser, parse_link
-from servicelib.rest_responses import unwrap_envelope
-from simcore_service_webserver.login import APP_LOGIN_CONFIG
+from simcore_service_webserver.login.cfg import APP_LOGIN_CONFIG
 from simcore_service_webserver.statics import INDEX_RESOURCE_NAME
 
 NEW_EMAIL = "new@mail.com"
