+++ services/web/server/tests/unit/with_dbs/fast/test_change_password.py
@@ -3,12 +3,10 @@
 # pylint:disable=redefined-outer-name
 
 from aiohttp import web
-from yarl import URL
 
 from pytest_simcore.helpers.utils_assert import assert_status
-from pytest_simcore.helpers.utils_login import LoggedUser, parse_link
-from servicelib.rest_responses import unwrap_envelope
-from simcore_service_webserver.login import APP_LOGIN_CONFIG
+from pytest_simcore.helpers.utils_login import LoggedUser
+from simcore_service_webserver.login.cfg import APP_LOGIN_CONFIG
 
 NEW_PASSWORD = "NewPassword1*&^"
