+++ services/web/server/tests/unit/with_dbs/fast/test_reset_password.py
@@ -8,9 +8,8 @@ from yarl import URL
 
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import NewUser, parse_link, parse_test_marks
-from servicelib.rest_responses import unwrap_envelope
 from simcore_service_webserver.db_models import ConfirmationAction, UserStatus
-from simcore_service_webserver.login import APP_LOGIN_CONFIG
+from simcore_service_webserver.login.cfg import APP_LOGIN_CONFIG
 from simcore_service_webserver.login.utils import get_random_string
 
 EMAIL, PASSWORD = "tester@test.com", "password"
